% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_K_svt.R
\name{choose_K_svt}
\alias{choose_K_svt}
\title{Implement SVT to choose K given symmetric adjacency matrix}
\usage{
choose_K_svt(A)
}
\arguments{
\item{A}{A symmetric adjacency matrix with entries of 0 or 1}
}
\description{
Function to estimate number of communities K from an adjacency matrix A
by computing the number of singular values above sqrt(n), where n is the number
of nodes in the graph.
}
\examples{
choose_K_svt(A)
}
\keyword{SBM,}
\keyword{SVT,}
