% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels.R
\name{get_labels}
\alias{get_labels}
\title{The get_labels function}
\usage{
get_labels(fit)
}
\arguments{
\item{fit}{A model object obtained with sbmlogit.mcmc()}
}
\description{
function to obtain remapped posterior centroid estimator of cluster labels
see Peng & Car. 2016.
}
\examples{
get_labels(fitK2)
}
\keyword{SBM}
