% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sbmlogit.R
\name{plot_sbmlogit}
\alias{plot_sbmlogit}
\title{The plot_sbmlogit function}
\usage{
plot_sbmlogit(fit, ground = NULL, labels = NULL, alpha = 0.5)
}
\arguments{
\item{fit}{A model object obtained with sbmlogit.mcmc()}

\item{ground}{An optional parameter specifying the variable name of the ground truth cluster labels in the original igraph data object}

\item{labels}{An optional parameter specifying the labels of each node to be used. If NULL, each node will be given a unique id label.}

\item{alpha}{Transparency of the edge color}
}
\description{
Function to plot data graph along with inferred cluster labels
}
\examples{
plot_sbmlogit(fitK2)
}
\keyword{SBM}
