% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sbmlogit.R
\name{plot_sbmlogit}
\alias{plot_sbmlogit}
\title{The plot_sbmlogit function}
\usage{
plot_sbmlogit(fit, ground = NULL)
}
\arguments{
\item{fit}{A model object obtained with sbmlogit.mcmc()}

\item{ground}{An optional parameter specifying the variable name of the ground truth cluster labels in the original igraph data object}
}
\description{
Function to plot data graph along with inferred cluster labels
}
\examples{
plot_sbmlogit(fitK2)
}
\keyword{SBM}
